// src/conversation/gateways/path-types.ts
export enum PathTypes {
  GET_CONVERSATION = 'get-conversation',
  JOIN_ROOM = 'join-room',
  LEAVE_ROOM = 'leave-room',
  SEND_MESSAGE = 'send-message',
  DELETE_MESSAGE = 'delete-message',
  ADD_TYPING = 'add-typing',
  REMOVE_TYPING = 'remove-typing',
  GET_TYPING = 'get-typing',
  UPDATE_CONVERSATIONS = 'update-conversations'
}